move store methods to mixins
