+++ puzzle/plugins/sql/mixins/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+from .actions import ActionsMixin
+from .case import CaseMixin
+from .variant import VariantMixin
