+++ puzzle/plugins/sql/mixins/actions/core.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+from .phenotype import PhenotypeActions
+from .gemini import GeminiActions
+from .genelist import GeneListActions
+from .resource import ResourceActions
+
+
+class ActionsMixin(PhenotypeActions, GeminiActions, GeneListActions,
+                   ResourceActions):
+    pass
